Ported against Mac OSX 10.2   autoconf related files     Makefile.* configure.* aclocal.m4   replacement file for hsearch() because of bug in Mac OSX     hsearch.h hsearch.c   added #if defined(__darwin__) among other BSD     rtptrans.c
Akira Tsukamoto

+++ hsearch.h
@@ -49,7 +49,7 @@ typedef enum {
         FIND, ENTER
 } ACTION;
 
-int      hcreate __P((unsigned int));
+int      hcreate __P((size_t));
 void     hdestroy __P((void));
 ENTRY   *hsearch __P((ENTRY, ACTION));
